requirement.txt
pandas==2.3.1
python-liquid==2.0.2
num2words==0.5.14